Schemas RPC  (#58)
* sync two branches

+++ common/primitives/src/schema.rs
@@ -1,1 +1,1 @@
-pub type SchemaId = u32;
+pub type SchemaId = u16;
